+++ playlist_bridge/__init__.py
@@ -1,3 +1,5 @@
 """Playlist Bridge package."""
 
-__version__ = "2.0.0-beta.3"
+__version__ = "2.0.0-beta.4"
+
+__build__ = "20260908.8"